Created AdminCourseList and StudentCourseList to inherit from CourseList
