+++ app/assets/javascripts/students/components/StudentDashboard.js
@@ -4,13 +4,13 @@ import { Link } from 'react-router'
 import { getUser } from '../../utils/user_helpers'
 import { RailsRoutes, ReactRoutes } from '../../shared/routes'
 
-import CourseList from './CourseList'
+import StudentCourseList from './StudentCourseList'
 
 class StudentDashboard extends React.Component {
   render() {
     return (
       <div>
-        <CourseList />
+        <StudentCourseList route={ReactRoutes.courseOutlinePath(value.id)}/>
       </div>
     )
   }
